Add files via upload
Dr. Pawaskar,

I've uploaded the datasets that we're planning to  use at this time, and a general outline for the project (Group Project -- Request for Proposal.docx).  Included the plan and responsibility assignment (Plan for Group Project - Immigration.xlsx).

Both of these documents are waiting for feedback but it's a summarized view of our meetings and what we've found so far. 

-Brandy